updated automated testing

+++ tests/test_enhancement.py
@@ -18,8 +18,11 @@ def handle_remove_readonly(func, path, exc):
         raise
 
 
-def clean_up_temp():
-    shutil.rmtree('./temp', ignore_errors=False, onerror=handle_remove_readonly)
+@pytest.fixture(autouse=True)
+def run_around_tests():
+    yield
+    if os.path.exists('./temp'):
+        shutil.rmtree('./temp', ignore_errors=False, onerror=handle_remove_readonly)
 
 
 def load_data_files():
@@ -54,26 +57,23 @@ def test_enhancement_validity():
         # print("control:         ", control_data_frame.iloc[0][data_element.variable_name])
         assert enhanced_data_frame.iloc[0][data_element.variable_name] == \
             control_data_frame.iloc[0][data_element.variable_name]
-    clean_up_temp()
 
 
 def test_geocodable_address():
-    file_path = 'validation/addresses-us-all.csv'
+    file_path = './validation/addresses-us-all.csv'
     data_key = main.get_data_key(file_path)
     input_data_frame = importer.import_file(file_path)
     input_data_frame = input_data_frame.iloc[50:52]
     input_data_frame.index = [0, 1]
     input_data_frame = geocoder.geocode_addresses_in_data_frame(input_data_frame, data_key)
     assert input_data_frame.iloc[0][constant.GEO_ID_NAME] == "04013618000"
-    clean_up_temp()
 
 
 def test_non_geocodable_address():
-    file_path = 'validation/addresses-us-all.csv'
+    file_path = './validation/addresses-us-all.csv'
     data_key = main.get_data_key(file_path)
     input_data_frame = importer.import_file(file_path)
     input_data_frame = input_data_frame.iloc[50:52]
     input_data_frame.index = [0, 1]
     input_data_frame = geocoder.geocode_addresses_in_data_frame(input_data_frame, data_key)
     assert input_data_frame.iloc[1][constant.GEO_ID_NAME] == constant.ADDRESS_NOT_GEOCODABLE
-    clean_up_temp()
